New implementation - author in ES6, share via esm
