+++ .npmignore
@@ -30,5 +30,3 @@
 /bower.json.ember-try
 /package.json.ember-try
 /test-apps
-
-/addon-docs-shared/node_modules
