--- addon-docs-shared/compile-markdown.js
@@ -1,150 +1,0 @@
-'use strict';
-
-const marked = require('marked');
-
-const hljs = require('highlight.js/lib/highlight');
-
-// Installed languages
-const javascript = require('highlight.js/lib/languages/javascript');
-const css = require('highlight.js/lib/languages/css');
-const handlebars = require('highlight.js/lib/languages/handlebars');
-const htmlbars = require('highlight.js/lib/languages/htmlbars');
-const json = require('highlight.js/lib/languages/json');
-const xml = require('highlight.js/lib/languages/xml');
-const diff = require('highlight.js/lib/languages/diff');
-const shell = require('highlight.js/lib/languages/shell');
-
-hljs.registerLanguage('javascript', javascript);
-hljs.registerLanguage('css', css);
-hljs.registerLanguage('handlebars', handlebars);
-hljs.registerLanguage('htmlbars', htmlbars);
-hljs.registerLanguage('json', json);
-hljs.registerLanguage('xml', xml);
-hljs.registerLanguage('diff', diff);
-hljs.registerLanguage('shell', shell);
-
-function highlightCode(code, lang) {
-  return hljs.getLanguage(lang) ? hljs.highlight(lang, code).value : code
-}
-
-module.exports = function compileMarkdown(source, config) {
-  let tokens = marked.lexer(source);
-  let markedOptions = {
-    highlight: highlightCode,
-    renderer: new HBSRenderer(config)
-  };
-
-  if (config && config.targetHandlebars) {
-    tokens = compactParagraphs(tokens);
-  }
-
-  return `<div class="docs-md">${marked.parser(tokens, markedOptions).trim()}</div>`;
-};
-
-// Whitespace can imply paragraphs in Markdown, which can result
-// in interleaving between <p> tags and block component invocations,
-// so this scans the Marked tokens to turn things like this:
-//    <p>{{#my-component}}<p>
-//    <p>{{/my-component}}</p>
-// Into this:
-//    <p>{{#my-component}} {{/my-component}}</p>
-function compactParagraphs(tokens) {
-  let compacted = [];
-
-  compacted.links = tokens.links;
-
-  let balance = 0;
-  for (let token of tokens) {
-    if (balance === 0) {
-      compacted.push(token);
-    } else if (token.text) {
-      let last = compacted[compacted.length - 1];
-      last.text = `${last.text} ${token.text}`;
-    }
-
-    let tokenText = token.text || '';
-    let textWithoutCode = tokenText.replace(/`[\s\S]*?`/g, '');
-
-    balance += count(/{{#/g, textWithoutCode);
-    balance += count(/<[A-Z]/g, textWithoutCode);
-    balance -= count(/{{\//g, textWithoutCode);
-    balance -= count(/<\/[A-Z]/g, textWithoutCode);
-  }
-
-  return compacted;
-}
-
-function count(regex, string) {
-  let total = 0;
-  while (regex.exec(string)) total++;
-  return total;
-}
-
-class HBSRenderer extends marked.Renderer {
-  constructor(config) {
-    super();
-    this.config = config || {};
-  }
-
-  codespan() {
-    return this._processCode(super.codespan.apply(this, arguments));
-  }
-
-  code() {
-    let code = this._processCode(super.code.apply(this, arguments));
-
-    return code.replace(/^<pre>/, '<pre class="docs-md__code">');
-  }
-
-  // Unescape markdown escaping in general, since it can interfere with
-  // Handlebars templating
-  text() {
-    let text = super.text.apply(this, arguments);
-    if (this.config.targetHandlebars) {
-      text = text
-        .replace(/&amp;/g, '&')
-        .replace(/&lt;/g, '<')
-        .replace(/&gt;/g, '>')
-        .replace(/&quot;|&#34;/g, '"')
-        .replace(/&apos;|&#39;/g, '\'');
-    }
-    return text;
-  }
-
-  // Escape curlies in code spans/blocks to avoid treating them as Handlebars
-  _processCode(string) {
-    if (this.config.targetHandlebars) {
-      string = this._escapeCurlies(string);
-    }
-
-    return string;
-  }
-
-  _escapeCurlies(string) {
-    return string
-      .replace(/{{/g, '&#123;&#123;')
-      .replace(/}}/g, '&#125;&#125;');
-  }
-
-  heading(text, level) {
-    let id = text.toLowerCase().replace(/<\/?.*?>/g, '').replace(/[^\w]+/g, '-');
-    let inner = level === 1 ? text : `<a href='#${id}' class='heading-anchor'>${text}</a>`;
-
-    return `
-      <h${level} id='${id}' class='docs-md__h${level}'>${inner}</h${level}>
-    `;
-  }
-
-  hr() {
-    return `<hr class='docs-md__hr'>`;
-  }
-
-  blockquote(text) {
-    return `<blockquote class='docs-md__blockquote'>${text}</blockquote>`;
-  }
-
-  link(href, title, text) {
-    const titleAttribute = title ? `title="${title}"` : '';
-    return `<a href="${href}" ${titleAttribute} class="docs-md__a">${text}</a>`;
-  }
-}
