+++ lib/broccoli/docs-compiler.js
@@ -7,7 +7,7 @@ const CachingWriter = require('broccoli-caching-writer');
 
 const Serializer = require('../serializers/main');
 
-const compileMarkdown = require('addon-docs-shared//compile-markdown');
+const compileMarkdown = require('../utils/compile-markdown');
 const NavigationIndexGenerator = require('./docs-compiler/navigation-index-generator');
 
 function compileDescriptions(objects) {
