+++ lib/preprocessors/markdown-template-compiler.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const stew = require('broccoli-stew');
-const compileMarkdown = require('addon-docs-shared/compile-markdown');
+const compileMarkdown = require('../utils/compile-markdown');
 
 module.exports = class MarkdownTemplateCompiler {
   constructor() {
