+++ lib/utils/compile-doc-descriptions.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const compileMarkdown = require('addon-docs-shared/compile-markdown');
+const compileMarkdown = require('./compile-markdown');
 
 module.exports = function compileDescriptions(docs) {
   for (let doc of docs) {
