+++ lib/utils/compile-markdown.js
@@ -1,0 +1,2 @@
+const esmRequire = require("esm")(module, { cjs: true });
+module.exports = esmRequire('../../addon/utils/compile-markdown').default;
